SewageMapAI/backend/requirements.txt
# Flask and related packages
flask==2.2.3
flask-cors==4.0.0
werkzeug==2.2.3
gunicorn==20.1.0

# Data handling and processing
numpy==1.24.3
pandas==2.0.0
joblib==1.2.0
opencv-python==4.7.0.72
pillow==9.5.0

# Database
sqlalchemy==2.0.9

# Utilities
python-dotenv==1.0.0
requests==2.28.2

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Machine Learning (optional)
scikit-learn==1.2.2
matplotlib==3.6.3

# Deep Learning for image segmentation
tensorflow==2.13.0
keras==2.13.1
albumentations==1.3.1
segmentation-models==1.0.1
kaggle==1.5.16
